+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "ihnatiev-site-v0.6.23b-r1";
+const CACHE_NAME = "ihnatiev-site-v0.6.23c-r1";
 const APP_SHELL = [
   "./",
   "./index.html",
@@ -26,6 +26,7 @@ const APP_SHELL = [
   "./config.js",
   "./files/content/home.json",
   "./files/content/activities.json",
+  "./files/content/pages.json",
   "./js/site-utils.js",
   "./js/gallery-lightbox.js",
   "./js/gallery-renderer.js",